Refactor secret-based auth flows
